+++ src/components/app.js
@@ -1,18 +1,48 @@
 import { Component } from 'preact';
 import { Router } from 'preact-router';
+import firebase from 'firebase/app';
 
 // Import Routes
 import Login from '../routes/login';
-import Account from '../routes/account/index';
 import Home from '../components/home';
+
+// Import Firebase Login
+import fireApp from '../base2';
+
+require('firebase/auth');
 
 export default class App extends Component {
   constructor() {
     super();
 
+    this.logout = this.logout.bind(this);
+    this.authenticate = this.authenticate.bind(this);
+
     this.state = {
       uid: 'hl8tgg53mkQIUmIh6D8SUsReTGD2',
+      email: 'user@uva.nl',
+      // email: null,
+      // uid: null,
     };
+  }
+
+  logout(e) {
+    e.preventDefault();
+    firebase.auth(fireApp).signOut();
+    this.setState({ uid: null, email: null });
+  }
+
+  authenticate(e, email, password) {
+    e.preventDefault();
+    fireApp
+      .auth()
+      .signInAndRetrieveDataWithEmailAndPassword(email, password)
+      .then(user => {
+        this.setState({
+          uid: user.user.uid,
+          email: user.user.email,
+        });
+      });
   }
 
   handleRoute = e => {
@@ -21,7 +51,7 @@ export default class App extends Component {
 
   render() {
     if (this.state.uid === null) {
-      return <Login />;
+      return <Login authenticate={this.authenticate} />;
     }
     return (
       <div id="app">
@@ -31,7 +61,12 @@ export default class App extends Component {
           <Home path="/add" uid={this.state.uid} />
           <Home path="/addMedia" uid={this.state.uid} />
           <Home path="/media/:mediaId" uid={this.state.uid} />
-          <Account path="/account" />
+          <Home
+            path="/account"
+            uid={this.state.uid}
+            email={this.state.email}
+            logout={this.logout}
+          />
         </Router>
       </div>
     );