+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.3] - 2018-01-17
+### Added
+- Login Component.
+- Account Component.
+
+### Changed
+- The InputGroup Component now accepts autocomplete as a prop.
+
 ## [0.0.2] - 2018-01-17
 ### Added
 - RemoveButton Component.